backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "chromadb>=1.5.7",
    "fastapi>=0.135.3",
    "google-auth>=2.38.0",
    "httpx>=0.28.1",
    "openai>=2.31.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.10.0",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]
